Add exception hint about str types when deserialize fails (#11)
The `message` parameter of the `deserialize` method takes `bytes`, as it
expects an Avro binary wrapped in Confluent Registry framing.

While integrating between Python and PHP services with CloudEvents, we
were inadvertently getting a `str` out of JSON decoding and feeding that
straight through, which led to a cryptic
`avrocurio.exceptions.InvalidWireFormatError: Invalid magic byte.
Expected 0, got '\x00'`

(Actually, lack of repr() on the expected value directly printed null
bytes that just looked like empty space, so it was even more unclear).

One would expect type checking to catch this, but `str` satisfies the
`bytes` type due to duck typing so type checkers don't complain about
a. This masks issues where the Avro bytes have been inadvertently
decoded and now get passed on as unicode code points.

The `Exception.add_note()` method offers a nice way to inject a hint to
users when this happens, without introducing the cost of runtype type
comparisons for each and every call to `deserialize()`.

This does require bumping the minimum supported Python version to 3.11,
but as that was released almost 3 years ago already, this feels like an
acceptable decision.